First version of LocationSnapshotReport
New report module: LocationSnapshotReport
New report superclass: LocationSnapshotReport
Refactored EnergyLevelReport to also use the new superclass
